+++ apps/client/src/common/components/playback-icon/PlaybackIcon.tsx
